+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+##
+- Bugfix disable track output when restarting ([#134](https://github.com/OpenRemise/Firmware/issues/134))
+
 ## 0.7.0
 - Add Ethernet support through [W5500](https://wiznet.io/products/ethernet-chips/w5500) ([#123](https://github.com/OpenRemise/Firmware/issues/123))
 - Add display support ([#116](https://github.com/OpenRemise/Firmware/issues/116))