+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
-##
+## 0.7.1
+- Add (supported) Z21 settings ([#138](https://github.com/OpenRemise/Firmware/issues/138))
 - Bugfix disable track output when restarting ([#134](https://github.com/OpenRemise/Firmware/issues/134))
 - Bugfix settings reset must always work ([#135](https://github.com/OpenRemise/Firmware/issues/135))
 - Bugfix Z21 `LAN_X_SET_STOP` and `LAN_X_SET_LOCO_E_STOP` ([#136](https://github.com/OpenRemise/Firmware/issues/136))